feat(skeleton_id): move hw accel meta to dedicated module
