+++ elasticai/creator/vhdl/skeleton_id.py
@@ -5,11 +5,7 @@ The skeleton id should be computed and set after all other code is generated.
 import logging
 from collections.abc import Iterable
 from hashlib import blake2b, file_digest
-from importlib.metadata import version as _pypackage_version
 from pathlib import Path
-
-import tomlkit as toml
-from tomlkit.toml_file import TOMLFile
 
 
 def compute_skeleton_id_hash(files: Iterable[Path]) -> bytes:
@@ -115,38 +111,3 @@ def _get_skeleton_pkg_file(build_dir: Path) -> Path:
         if file.name == "skeleton_pkg.vhd":
             return file
     raise IOError(f"could not find skeleton_pkg.vhd file in {build_dir}")
-
-
-class MetaFile:
-    version = "0.1"
-
-    def __init__(self, data):
-        self._data = data
-
-    @classmethod
-    def load(cls, file: Path) -> "MetaFile":
-        with open(file, "r") as f:
-            data = toml.load(f)
-        return cls(data)
-
-    def set_skeleton_id(self, skeleton_id: bytes):
-        self._data["skeleton_id"] = skeleton_id.hex()
-
-    def save(self, file: Path) -> None:
-        f = TOMLFile(file)
-        f.write(self._data)
-
-    @classmethod
-    def default(cls) -> "MetaFile":
-        creator_version = _pypackage_version("elasticai.creator")
-        data = dict(
-            meta_version=cls.version,
-            creator_version=creator_version,
-            hw_accelerator_version="0.1",
-            skeleton_id='"<invalid>"  # this should be computed and set'
-            " automatically",
-            name="<your_network>",
-            description="""An optional possibly very long description of what
-                            your accelerator does and how to use it.""",
-        )
-        return cls(data)
